data: Add network tasks + de/serialize from/to JSON
Work in progress. Basic idea at the moment is to convert local tasks to
network tasks, then serialize that list into a JSON string, and send
that string to the backend to be stored.

TODO: we're using / modifying a lot of code from the Android tutorials
and architecture sample, so we should include the required licensing.