ui: Set up login screen to work with a signup screen
"Sign up" button on the login screen will take the user to a signup
screen. When they successfully register, they will be taken back to the
login screen. We add a `messageFromSignup` argument so that external
files can update the snackbar on the login screen. We can't use a
snackbar from the signup screen because we navigate away from the signup
screen, so such a snackbar would not be displayed.
We could also have gone for a global snackbarHostState that both the
login screen and view model have access to; but that's more complicated
and can be an improvement in the future.