ui: Set up navigation to signup screen
We handle the logic mentioned in a2fed76 in the nav host. We keep a
mutable state in the nav host composable that gets updated when a sign
up is successful. The message is sent to the login screen, and then the
value is set to null.
The code is fragile, and as mentioned a2fed76, a snackbarHostState that
can be accessed by both login and signup screens would probably be
better.