ui: Fix bug where lazy list always scrolls to bottom of list
Using LaunchedEffect(taskList) will always scroll to the bottom of the
list because taskList is always empty upon first composition. The tasks
come from a Flow, so the taskList gets populated after composition
already happens.

This solution is rough, we have to use a private var to inside the
ViewModel and a Boolean in the UiState to keep track of whether we need
to scroll, but it works, so whatever.