ui: Fix double click back button showing blank screen
We weren't checking the Lifecycle of the NavBackStackEntry to verify if
the screen is resumed. So when double clicking the back button, we're
popping two screens off the stack. In most cases, that means there's
nothing left to display, so we get a blank screen.

See https://github.com/google/accompanist/issues/1408